custom serial: pass device_data_t to bluetooth operation function
That odd 'get_qt_serial_ops()' helper function originally existed
because of build issues, but it turns out that we will actually want to
use different custom serial operations for different kinds of bluetooth
devices, rather than always forcing the RFCOMM model.

So to prepare for that, pass in the 'device_data_t' to the op function
helper, so that it can look up the device details and give different
serial operation structures for different kinds of devices.

This commit doesn't actually change any existing semantics, it just
paves the way for other serial communication emulation.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>