Expand 'bluetooth_mode' from a boolean to an emun
We don't want to say just "bluetooth or not".  We want to enumerate the
different bluetooth modes, so that we have RFCOMM vs BLE vs possible
other cases.

Leave the "BT_NONE" mode as zero, so that you can continue to do

	if (data->bluetooth_mode) ...

to check whether we're supposed to use an bluetooth connection.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>